chore: update service paths
