+++ app-data.json
@@ -1,6 +1,6 @@
 {
   "name": "rolnopol-app",
-  "version": "1.45.0",
+  "version": "1.46.0",
   "description": "Simple application prepared and developed only for testing purposes",
   "keywords": [
     "json-server",
