+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rolnopol-app",
-  "version": "1.45.0",
+  "version": "1.46.0",
   "description": "A Rolnopol app with JSON database, authentication, and admin panel",
   "main": "api/index.js",
   "scripts": {
@@ -10,12 +10,12 @@
     "start": "node ./api/index.js",
     "start-debug": "cross-env DEBUG_MODE=true node ./api/index.js",
     "dev": "node ./api/index.js",
-    "greenhouse": "node ./grpc/greenhouse-server/index.js",
-    "greenhouse:health": "node ./grpc/client-health.js",
-    "greenhouse:plant": "node ./grpc/client-plant.js",
+    "greenhouse": "node ./external-services/greenhouse/greenhouse-server/index.js",
+    "greenhouse:health": "node ./external-services/greenhouse/client-health.js",
+    "greenhouse:plant": "node ./external-services/greenhouse/client-plant.js",
     "greenhouse:test": "cross-env NODE_ENV=test vitest run tests/unit/greenhouse.service.test.js tests/unit/greenhouse.tick-engine.test.js && cross-env NODE_ENV=test vitest run tests/greenhouse-grpc.test.js tests/greenhouse-rest.test.js tests/greenhouse-pages-gating.test.js",
-    "tasklab": "node ./grpc/tasklab-server/index.js",
-    "tasklab:demo": "node ./grpc/client-tasklab.js",
+    "tasklab": "node ./external-services/tasklab/tasklab-server/index.js",
+    "tasklab:demo": "node ./external-services/tasklab/client-tasklab.js",
     "tasklab:test": "cross-env NODE_ENV=test vitest run tests/unit/tasklab.service.test.js && cross-env NODE_ENV=test vitest run tests/tasklab-grpc.test.js tests/tasklab-rest.test.js tests/tasklab-pages-gating.test.js",
     "format": "npx prettier --write .",
     "bump-version": "node bump-version.js minor"
