+++ public/schema/openapi.json
@@ -2,7 +2,7 @@
   "openapi": "3.0.0",
   "info": {
     "title": "Rolnopol",
-    "version": "1.45.0",
+    "version": "1.46.0",
     "description": "API documentation for the Rolnopol service with versioning support",
     "contact": {
       "name": "API Support",
